requirements.txt
leafmap==0.42.7
leafmap==0.57.5
matplotlib==3.10.7
numpy==2.3.4
pandas==2.3.3
Pillow==12.0.0
plotly==5.24.1
rasterio==1.4.3
streamlit==1.51.0
